historias_da_bola/requirements.txt
django==1.3.1
django-mptt==0.5.1
django-cms
django-reversion
django-tinymce
PIL
